Restructure hints server
